refactor: coupopn 세부분리
